add incremental utility functions (from delta fixes, updates)
Added:
- IsLocalWindows() (consolidate from multiple places)
- ToLocal(), ToFileLocal() - for converting between file:/ and / style absolute paths
- HasParentage() to check whether one path is some kind ancestor of another

Also:
- moved cpp unit tests from test_file_system to test_path.cpp
