+++ test/common/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library_unity(
   test_cast.cpp
   test_checksum.cpp
   test_file_system.cpp
+  test_path.cpp
   test_caching_file_system_wrapper.cpp
   test_hyperlog.cpp
   test_lru_cache.cpp
